refactor(abs): split importer domain helpers
